+++ frontend/src/views/transfer/index.css
@@ -1,69 +1,4 @@
-.iam-transfer-group-perm-wrapper {
-    padding-bottom: 76px;
-    .iam-perm-item {
-        background: #fff;
-        border-radius: 2px;
-        border: 1px solid #fff;
-        box-shadow: 0px 1px 2px 0px rgba(49, 50, 56, .1);
-        &:hover {
-            box-shadow: 0px 2px 4px 0px rgba(49, 50, 56, .15);
-        }
-        &.is-not-expanded:hover {
-            .header {
-                cursor: pointer;
-            }
-        }
-        .header {
-            display: flex;
-            justify-content: flex-start;
-            position: relative;
-            padding: 0 30px;
-            height: 40px;
-            line-height: 40px;
-            font-size: 14px;
-            color: #63656e;
-            border-radius: 2px;
-            cursor: pointer;
-            .expanded-icon {
-                position: absolute;
-                top: 13px;
-                left: 10px;
-            }
-            .title {
-                font-weight: 500;
-                cursor: pointer;
-                color: #313238;
-            }
-            .sub-title {
-                margin-left: 17px;
-                color: rgba(0, 0, 0, 0.65);
-                .reason {
-                    font-size: 12px;
-                    color: #979ba5;
-                }
-            }
-        }
-        .content {
-            position: relative;
-            .expand-action {
-                padding: 0 0 5px 0;
-                margin-top: -10px;
-                width: 100%;
-                line-height: 16px;
-                font-size: 12px;
-                color: #979ba5;
-                text-align: center;
-                cursor: pointer;
-                i {
-                    font-size: 16px;
-                }
-            }
-            .slot-content {
-                padding: 0 30px 20px 30px;
-                min-height: 60px;
-            }
-        }
-    }
+.iam-transfer-wrapper {
     .fixed-action {
         position: fixed;
         left: 0;
